services/storage/requirements.txt
psycopg2-binary>=2.9.0
redis>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0

